refactor(browser): split 2979-line browser_tool.py into focused modules
Extract monolithic browser_tool.py into 7 focused modules:
- browser_utils.py (437 lines) — PATH, config, env detection
- browser_state.py (23 lines) — shared mutable state
- browser_session.py (554 lines) — session lifecycle, cleanup, reaping
- browser_local.py (423 lines) — local CLI backend, content extraction
- browser_actions.py (619 lines) — core actions (navigate, click, type, etc.)
- browser_actions_media.py (313 lines) — screenshots, images, vision
- browser_actions_extra.py (140 lines) — hover, highlight, drag, select, etc.
- browser_tool.py (653 lines) — schemas + registry (was 2979 lines)

All existing imports keep working via re-exports from browser_tool.py.
